remove old UserAvatarView and update settings screen
