+++ packages/apolloschurchapp/src/user-settings/index.js
@@ -17,14 +17,9 @@ import {
   ActivityIndicator,
 } from '@apollosproject/ui-kit';
 import { WebBrowserConsumer } from 'apolloschurchapp/src/ui/WebBrowser';
-import AvatarForm from 'apolloschurchapp/src/ui/UserAvatarView/AvatarForm';
 
 import { GET_LOGIN_STATE, LOGOUT } from '@apollosproject/ui-auth';
-
-const AvatarView = styled({
-  alignItems: 'center',
-  justifyContent: 'center',
-})(PaddedView);
+import ChangeAvatar from './ChangeAvatar';
 
 class UserSettings extends PureComponent {
   static navigationOptions = () => ({
@@ -47,9 +42,8 @@ class UserSettings extends PureComponent {
           return (
             <BackgroundView>
               <ScrollView>
-                <AvatarView>
-                  <AvatarForm text />
-                </AvatarView>
+                <ChangeAvatar />
+
                 <WebBrowserConsumer>
                   {(openUrl) => (
                     <>
