+++ packages/apolloschurchapp/src/user-settings/uploadPhoto.js
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag';
 import ImagePicker from 'react-native-image-picker';
 import { ReactNativeFile } from 'apollo-upload-client';
-import GET_USER_PROFILE from '../../tabs/connect/getUserProfile';
+import GET_USER_PROFILE from './getUserPhoto';
 
 const options = {
   title: 'Select Profile Image',
